+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Keep focus on keyboard-moved phrase.
 - Render Markup values as symbols for debugger to highlight.
 - Fixed Phrase dragging in output view.
+- Performance improvements to editor.
 
 ### Maintenance